more work on effects
changed DNAData to store effect as a list rather than an array. This
should more easily allow effects to be added to the data on the fly.

Also worked more on the general effect management and registered the
even handlers. However, no effects are tested so far so it is hard to
tell if the system works. I will still need to do some base edits to the
Entity Living class to better work with the Entities that have effects.